Remove build.hpp from public API
