+++ tt_metal/api/tt-metalium/device.hpp
@@ -12,7 +12,6 @@
 #include "work_executor_types.hpp"
 #include "data_types.hpp"
 #include "program_device_map.hpp"
-#include "build.hpp"
 #include "hal.hpp"
 #include "command_queue_interface.hpp"
 #include "sub_device_types.hpp"
