+++ tt_metal/api/tt-metalium/device_impl.hpp
@@ -9,10 +9,10 @@
 
 #include "device.hpp"
 #include "hostdevcommon/common_values.hpp"
+#include "hostdevcommon/kernel_structs.h"  // Leaked up to ttnn level from here
 #include "work_executor_types.hpp"
 #include "data_types.hpp"
 #include "program_device_map.hpp"
-#include "build.hpp"
 #include "hal.hpp"
 #include "command_queue_interface.hpp"
 #include "command_queue.hpp"
