+++ tt_metal/jit_build/CMakeLists.txt
@@ -4,7 +4,7 @@ set(JIT_BUILD_SRCS
     ${CMAKE_CURRENT_SOURCE_DIR}/data_format.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/genfiles.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/kernel_args.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/settings.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/jit_build_options.cpp
 )
 
 add_library(jit_build OBJECT ${JIT_BUILD_SRCS})
