+++ tt_metal/jit_build/build.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <build.hpp>
+#include "build.hpp"
 
 #include <filesystem>
 #include <fstream>
